Script py3 and no bias update
As this require python 3 and based on others, it is suggested to add
this.

Also there is no update to bias which some other implementation and
course may want it.  Just a note.  Not sure I dare to do the update
bias in case there is a mistake.  At least not in this pull request.